Fonts look much nicer, similar to how they were
Set a fragment's alpha based on the color of the rendered glyph
fragment. This keeps the anti-aliased font rendering working
even with transparency enabled.

There's still a strange issue if you switch windows and return.
Ghosts of the on-screen glyphs remain in the background after they
should no longer be drawn. Resizing the window fixes the problem...